uptime-kuma/pyproject.toml
[project]
name = "uptime-kuma"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.2.1",
    "uptime-kuma-api>=1.2.1",
]
